Use the texture array render targets from Godot
Updated to Tilt Five 1.4.0-beta.6 NDK
No longer needs the custom copy_texture function to the added to Godot.
No longer needs patched Godot
