+++ SConstruct
@@ -2,9 +2,7 @@
 import os
 import sys
 
-tilt_five_headers_path = 'extension/TiltFiveNDK/include/include'
-tilt_five_library_path = 'extension/TiltFiveNDK/lib/win64'
-tilt_five_library = 'TiltFiveNative.dll.if'
+
 
 VariantDir('build/src','extension/src', duplicate=False)
 VariantDir('build/T5Integration','extension/T5Integration', duplicate=False)
@@ -12,6 +10,9 @@ VariantDir('build/T5Integration','extension/T5Integration', duplicate=False)
 env = SConscript('godot-cpp/SConstruct')
 env['CXXFLAGS'].remove('/std:c++17')
 env.Append(CXXFLAGS=['/std:c++20'])
+tilt_five_headers_path = 'extension/TiltFiveNDK/include'
+tilt_five_library_path = 'extension/TiltFiveNDK/lib/' + { 'windows' : 'win/x86_64', 'linux' : 'linux/x86_64'}[env["platform"]]
+tilt_five_library = 'TiltFiveNative.dll.if'
 
 # For the reference:
 # - CCFLAGS are compilation flags shared between C and C++
@@ -27,7 +28,7 @@ sources = Glob('build/src/*.cpp')
 sources += Glob('build/T5Integration/*.cpp')
 
 env.Append(LIBPATH=[tilt_five_library_path])
-env.Append(LIBS=[tilt_five_library, 'Opengl32'])
+env.Append(LIBS=[tilt_five_library])
 
 if env['platform'] == 'windows':
     env.Append(CXXFLAGS=['/Zc:__cplusplus'])
@@ -36,8 +37,8 @@ if env['platform'] == 'windows':
         source=sources,
     )
 
-f1 = env.Command('example/addons/tiltfive/libgdtiltfive{}{}'.format(env['suffix'], env['SHLIBSUFFIX']), library, Copy('$TARGET', '$SOURCE') )
-f2 = env.Command('example/addons/tiltfive/TiltFiveNative.dll', 'extension/TiltFiveNDK/lib/win64/TiltFiveNative.dll', Copy('$TARGET', '$SOURCE') )
+f1 = env.Command('example/addons/tiltfive/bin/libgdtiltfive{}{}'.format(env['suffix'], env['SHLIBSUFFIX']), library, Copy('$TARGET', '$SOURCE') )
+f2 = env.Command('example/addons/tiltfive/bin/TiltFiveNative.dll', tilt_five_library_path + '/TiltFiveNative.dll', Copy('$TARGET', '$SOURCE') )
 
 env.Alias('example', [f1, f2])
