+++ example/addons/tiltfive/tiltfive.gdextension
@@ -1,6 +1,7 @@
 [configuration]
 
 entry_symbol = "tiltfive_library_init"
+compatibility_minimum = "4.1.0"
 
 [libraries]
